+++ Source/Clockwork/Audio/OggVorbisSoundStream.cpp
@@ -1,3 +1,7 @@
+
+
+#include "../Precompiled.h"
+
 #include "../Audio/OggVorbisSoundStream.h"
 #include "../Audio/Sound.h"
 
@@ -43,7 +47,7 @@ unsigned OggVorbisSoundStream::GetData(signed char* dest, unsigned numBytes)
     stb_vorbis* vorbis = static_cast<stb_vorbis*>(decoder_);
 
     unsigned channels = stereo_ ? 2 : 1;
-    unsigned outSamples = stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)dest, numBytes >> 1);
+    unsigned outSamples = (unsigned)stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)dest, numBytes >> 1);
     unsigned outBytes = (outSamples * channels) << 1;
 
     // Rewind and retry if is looping and produced less output than should have
@@ -51,7 +55,8 @@ unsigned OggVorbisSoundStream::GetData(signed char* dest, unsigned numBytes)
     {
         numBytes -= outBytes;
         stb_vorbis_seek_start(vorbis);
-        outSamples = stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)(dest + outBytes), numBytes >> 1);
+        outSamples =
+            (unsigned)stb_vorbis_get_samples_short_interleaved(vorbis, channels, (short*)(dest + outBytes), numBytes >> 1);
         outBytes += (outSamples * channels) << 1;
     }
